requirements.txt
streamlit
crewai==0.28.8
groq
python-dotenv
langchain
langchain-groq
reportlab
chromadb